Continue to elf and multiboot for function call stack

+++ src/include/debug.h
@@ -19,7 +19,7 @@
 #define CHECK_FLAG(flags,bit)   ((flags) & (1 << (bit)))
 #define ELF32_ST_TYPE(i) ((i)&0xf)
 
-void debug_init(uint32_t magic, multiboot_info_t * mb);
+void debug_init(uint64_t magic, multiboot_info_t * mb);
 void print_registers(pt_regs_t * regs);
